also check negations in matrix vector test

+++ test.sh
@@ -1,4 +1,7 @@
-#!/bin/sh
-rm test.out
+#!/bin/bash
+
+if [ -f test.out ]; then
+    rm test.out
+fi
 g++ -march=native -O3 -std=c++11 *.cpp *.c -o test.out
 ./test.out
